move debug-console specific things to DebugConsoleAppender
